update: change radiobutton color for consistency
