+++ app/src/main/res/drawable/ic_baseline_radio_button_checked_28.xml
@@ -1,4 +1,4 @@
-<vector android:height="28dp" android:tint="#4CAF50"
+<vector android:height="28dp" android:tint="#536DFE"
     android:viewportHeight="24" android:viewportWidth="24"
     android:width="28dp" xmlns:android="http://schemas.android.com/apk/res/android">
     <path android:fillColor="@android:color/white" android:pathData="M12,7c-2.76,0 -5,2.24 -5,5s2.24,5 5,5 5,-2.24 5,-5 -2.24,-5 -5,-5zM12,2C6.48,2 2,6.48 2,12s4.48,10 10,10 10,-4.48 10,-10S17.52,2 12,2zM12,20c-4.42,0 -8,-3.58 -8,-8s3.58,-8 8,-8 8,3.58 8,8 -3.58,8 -8,8z"/>
